almost working

+++ pentabots/DualEncoder.hpp
@@ -69,8 +69,8 @@ public:
     volatile int8_t direction;
     float position = 0;
     uint16_t counts_per_revolution = 700;
-    volatile long l_count = 0;
-    volatile long r_count = 0;
+    volatile int16_t l_count = 0;
+    volatile int16_t r_count = 0;
     uint32_t prev_time;
     bool read = false;
